♻️ Refactor admin layout and remove redundant components
管理者機能のレイアウトを簡素化し、重複コンポーネントを削除
- AdminSidebar.tsx, BottomNavigation.tsx, ConditionalLayout.tsx を削除
- app/layout.tsx で統一レイアウトを採用
- Header.tsx に管理者ナビゲーションを統合
- フローティングヘッダーデザインを全ページで統一
