+++ app/layout.tsx
@@ -1,7 +1,8 @@
 import type { Metadata } from "next";
 import "./globals.css";
-import ConditionalLayout from "@/components/ConditionalLayout";
 import Background from "@/components/Background";
+import Header from "@/components/Header";
+import Footer from "@/components/Footer";
 
 export const metadata: Metadata = {
   title: "スキー・スノーボードスクール シフト管理システム",
@@ -13,7 +14,11 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
     <html lang="ja">
       <body className="antialiased">
         <Background />
-        <ConditionalLayout>{children}</ConditionalLayout>
+        <div className="min-h-screen flex flex-col">
+          <Header />
+          <main className="flex-1 pt-32">{children}</main>
+          <Footer />
+        </div>
       </body>
     </html>
   );
