--- components/ConditionalLayout.tsx
@@ -1,26 +1,0 @@
-"use client";
-
-import { usePathname } from "next/navigation";
-import Header from "./Header";
-import Footer from "./Footer";
-
-interface ConditionalLayoutProps {
-  children: React.ReactNode;
-}
-
-export default function ConditionalLayout({ children }: ConditionalLayoutProps) {
-  const pathname = usePathname();
-  const isAdminPage = pathname.startsWith("/admin");
-
-  if (isAdminPage) {
-    return <>{children}</>;
-  }
-
-  return (
-    <div className="min-h-screen flex flex-col">
-      <Header />
-      <main className="flex-1 sm:pt-20">{children}</main>
-      <Footer />
-    </div>
-  );
-}
